requirements.txt
discord.py==2.6.0
python-dotenv==1.1.1
requests==2.32.5
apscheduler==3.11.0